Model changed after meeting
