+++ ParcelDistributionCenter/ParcelDistributionCenter.Logic/CourierHandler.cs
@@ -1,4 +1,4 @@
-﻿using ParcelDistributionCenter.Model;
+﻿using ParcelDistributionCenter.Model.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
