+++ ParcelDistributionCenter/ParcelDistributionCenter.Model/Interfaces/IParcelHandler.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ParcelDistributionCenter.Model
+namespace ParcelDistributionCenter.Model.Interfaces
 {
     internal interface IParcelHandler
     {
