+++ ParcelDistributionCenter/ParcelDistributionCenter.Model/Models/Center.cs
@@ -1,10 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using ParcelDistributionCenter.Model.Interfaces;
 
-namespace ParcelDistributionCenter.Model
+namespace ParcelDistributionCenter.Model.Models
 {
     public class Center : IParcelHandler, IParcelPlace
     {
@@ -17,8 +13,7 @@ namespace ParcelDistributionCenter.Model
         public List<string> Parcels_numbers_list { get; set; }
         public List<string> Connected_lockers_ids { get; set; }
 
-
-        public Center(string id,string address, int big_lockers_count, int medium_lockers_count, int small_lockers_count) // konstruktor przy starcie programu
+        public Center(string id, string address, int big_lockers_count, int medium_lockers_count, int small_lockers_count) // konstruktor przy starcie programu
         {
             Id = id;
             Address = address;
@@ -30,7 +25,7 @@ namespace ParcelDistributionCenter.Model
             Connected_lockers_ids = new List<string>();
         }
 
-        public Center( string address, int big_lockers_count, int medium_lockers_count, int small_lockers_count) // konstruktor przy dodawaniu centrum
+        public Center(string address, int big_lockers_count, int medium_lockers_count, int small_lockers_count) // konstruktor przy dodawaniu centrum
         {
             Id = "C1"; // do opracowania auto nadawanie GUID
             Address = address;
@@ -38,17 +33,18 @@ namespace ParcelDistributionCenter.Model
             Big_lockers_count = big_lockers_count;
             Medium_lockers_count = medium_lockers_count;
             Small_lockers_count = small_lockers_count;
-            Parcels_numbers_list= new List<string>();
+            Parcels_numbers_list = new List<string>();
             Connected_lockers_ids = new List<string>();
         }
 
-        public void AddConnectedLocker( string Locker_Id) 
+        public void AddConnectedLocker(string Locker_Id)
         {
             if (!Connected_lockers_ids.Contains(Locker_Id))
             {
                 Connected_lockers_ids.Add(Locker_Id);
             }
         }
+
         public void AddConnectedLocker(List<string> Locker_Id)
         {
             foreach (string Id in Locker_Id)
@@ -59,6 +55,5 @@ namespace ParcelDistributionCenter.Model
                 }
             }
         }
-
     }
 }
