+++ ParcelDistributionCenter/ParcelDistributionCenter.Model/Models/Parcel.cs
@@ -1,6 +1,6 @@
 ﻿using ParcelDistributionCenter.Model.Enums;
 
-namespace ParcelDistributionCenter.Model
+namespace ParcelDistributionCenter.Model.Models
 {
     public class Parcel
     {
@@ -29,33 +29,9 @@ namespace ParcelDistributionCenter.Model
             Status = status;
             Current_location_id = current_location_id;
             Registered = registered;
-            Deliver_date = deliver_date; 
+            Deliver_date = deliver_date;
             Expire_date = null;
         }
-
-       /* public Parcel(ParcelSize size, string sender_email, string send_parcel_id, string recipient_email, string delivery_parcel_id) // konstruktor przy wprowadzaniu paczki
-        {
-           Number = "1"; // Do opracowania algorytm nadawania numerów paczek;
-            Size = size;
-            Sender_email = sender_email;
-            Send_parcel_id = send_parcel_id;
-            Recipient_email = recipient_email;
-            Delivery_parcel_id = delivery_parcel_id;
-            Status = (Status)0;
-            Current_location_id = "Sender";
-            Courier_id = "0";
-            Registered = DateTime.Now;
-            Expire_date = Registered.AddDays(2);
-        }
-
-        */
-
-
-        //public string Delivery_address { get; set; } -  tego parametru nie rozumiem - dostarczamy do paczkomatu więc wszystko jest już w Parcel_id.
-        // public string Sender_name { get; set; } - do klienta
-        // public string Recipient_name { get; set; } - do klienta
-        // public string Sender_phone { get; set; } do klienta
-        // public string Recipient_phone { get; set; }  do klienta
-        // public string Sender_address { get; set; }  do klienta
     }
-}
+         
+ }
